asteroid spacing much better

+++ Assets/Scripts/GameSettings.cs
@@ -21,12 +21,12 @@ public class GameSettings : ScriptableObject
     private float moveSpeed;
 
     [SerializeField]
-    private int gridLength;
+    private float gridLength;
 
     public float AsteroidSpeed { get => asteroidSpeed; set => asteroidSpeed = Mathf.Clamp(value, 0f, 1000f); }
     public float AsteroidSpawnInterval { get => asteroidSpawnInterval; set => asteroidSpawnInterval = Mathf.Clamp(value, 0f, 100f); }
     public int AsteroidsInGroup { get => asteroidsInGroup; set => asteroidsInGroup = Mathf.Clamp(value, 0, 9); }
     public float MoveDistance { get => moveDistance; set => moveDistance = value; }
     public float MoveSpeed { get => moveSpeed; set => moveSpeed = value; }
-    public int GridLength { get => gridLength; set => gridLength = value; }
+    public float GridLength { get => gridLength; set => gridLength = value; }
 }
